fix tensorflow_common

+++ src/frontends/onnx/frontend/src/utils/tensor_external_data.cpp
@@ -47,11 +47,11 @@ std::string TensorExternalData::load_external_data(const std::string& model_dir)
     if (external_data_stream.fail())
         throw error::invalid_external_data{*this};
 
-    uint64_t filesize = external_data_stream.tellg();
+    int filesize = external_data_stream.tellg();
     if (filesize == -1) {
         throw error::invalid_external_data{*this};
     }
-    if (m_offset + m_data_length > filesize) {
+    if (m_offset + m_data_length > (uint64_t)filesize) {
         throw error::invalid_external_data{*this};
     }
